разбила по компонентам
